requirements.txt
# TiltScope - Python Dependencies
# Cloud9 x JetBrains Hackathon 2025

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
